Relocate autocomplete shared dir
